+++ src/Shamsheer.Messenger.Api/Shamsheer.Messenger.Api.csproj
@@ -19,9 +19,9 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="wwwroot\Media\Files\" />
     <Folder Include="wwwroot\Media\Images\" />
     <Folder Include="wwwroot\Media\Audios\" />
-    <Folder Include="wwwroot\Media\Files\" />
     <Folder Include="wwwroot\Media\ProfilePictures\Groups\" />
     <Folder Include="wwwroot\Media\ProfilePictures\Channels\" />
     <Folder Include="wwwroot\Media\ProfilePictures\Users\" />
