+++ src/Shamsheer.Service/Interfaces/Commons/IFileService.cs
@@ -1,10 +1,12 @@
-﻿using System.Threading.Tasks;
+﻿using System.IO;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 
 namespace Shamsheer.Service.Interfaces.Commons;
 
 public interface IFileService
 {
-    Task<string> UploadImageAsync(IFormFile image, string rootpath, string filePath);
-    Task<bool> DeleteImageAsync(string subpath);
+    Task<string> UploadAsync(Stream file, string path);
+    Task<byte[]> DownloadAsync(string path);
+    Task<bool> DeleteAsync(string path);
 }
